[FIX]: 뷰홀더 class로뺌 innerClass->class로 사진보냈을때 채팅메시지에서 (사진)표시
