+++ app/src/main/java/com/example/frompet/ui/chat/adapter/ChatMessageAdapter.kt
@@ -22,24 +22,26 @@ import com.example.frompet.databinding.ItemOtherMessageBinding
 import com.example.frompet.data.model.ChatMessage
 import com.example.frompet.data.model.User
 import com.example.frompet.databinding.ItemDateHeaderBinding
+import com.example.frompet.ui.chat.viewmodel.MessageViewModel
+import com.example.frompet.ui.setting.fcm.FCMNotificationViewModel
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import java.util.Date
 import java.util.Locale
 
 
-class ChatMessageAdapter(var context: Context) :
+class ChatMessageAdapter(var context: Context,
+                         private val fcmViewModel: FCMNotificationViewModel,
+                         private val messageViewModel: MessageViewModel,
+                         private val user: User
+                         ) :
     ListAdapter<ChatItem, RecyclerView.ViewHolder>(DiffCallback()) {
-
-    private val firestore = FirebaseFirestore.getInstance()
-
     companion object {
         const val IMAGE_URL = "image_url"
         const val VIEW_TYPE_DATE_HEADER = 0
         const val VIEW_TYPE_MY_MESSAGE = 1
         const val VIEW_TYPE_OTHER_MESSAGE = 2
     }
-
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         val inflater = LayoutInflater.from(parent.context)
         return when (viewType) {
@@ -49,11 +51,11 @@ class ChatMessageAdapter(var context: Context) :
             }
             VIEW_TYPE_MY_MESSAGE -> {
                 val binding = ItemMyMessageBinding.inflate(inflater, parent, false)
-                MyMessageViewHolder(binding)
+                MyMessageViewHolder(binding,fcmViewModel,messageViewModel,user)
             }
             VIEW_TYPE_OTHER_MESSAGE -> {
                 val binding = ItemOtherMessageBinding.inflate(inflater, parent, false)
-                OtherMessageViewHolder(binding)
+                OtherMessageViewHolder(binding,context)
             }
             else -> throw IllegalArgumentException("Invalid view type")
         }
@@ -86,74 +88,7 @@ class ChatMessageAdapter(var context: Context) :
             }
         }
     }
-
-    inner class DateHeaderViewHolder(private val binding: ItemDateHeaderBinding) : RecyclerView.ViewHolder(binding.root) {
-        fun bind(date: String) {
-            binding.tvDateHeader.text = date
-        }
-    }
-
-
-    inner class MyMessageViewHolder(private val binding: ItemMyMessageBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-        fun bind(chatMessage: ChatMessage) = with(binding) {
-            if (chatMessage.imageUrl.isNullOrEmpty().not()) {
-                ivMessageImage.isVisible = true
-                tvMessage.isVisible = false
-                tvTime.text = formatTimeStamp(chatMessage.timestamp)
-                ivMessageImage.load(chatMessage.imageUrl) {
-                    error(R.drawable.kakaotalk_20230825_222509794_01)
-                }
-                clickListener(ivMessageImage, chatMessage.imageUrl)
-            } else {
-                ivMessageImage.isVisible = false
-                tvMessage.isVisible = true
-                tvMessage.setBackgroundResource(R.drawable.chat2)
-                tvMessage.text = chatMessage.message
-                tvTime.text =
-                    formatTimeStamp(chatMessage.timestamp)
-            }
-        }
-    }
-
-    inner class OtherMessageViewHolder(private val binding: ItemOtherMessageBinding) :
-        RecyclerView.ViewHolder(binding.root) {
-
-        fun bind(chatMessage: ChatMessage) = with(binding) {
-            if (chatMessage.imageUrl.isNullOrEmpty().not()) {
-                ivMessageImage.isVisible = true
-                tvMessage.isVisible = false
-                tvName.text = chatMessage.senderPetName
-                tvTime.text = formatTimeStamp(chatMessage.timestamp)
-                ivMessageImage.load(chatMessage.imageUrl) {
-                    error(R.drawable.kakaotalk_20230825_222509794_01)
-                }
-                clickListener(ivMessageImage, chatMessage.imageUrl)
-            } else {
-                ivMessageImage.isVisible = false
-                tvMessage.isVisible = true
-                tvName.text = chatMessage.senderPetName
-                tvMessage.setBackgroundResource(R.drawable.chat1)
-                tvMessage.text = chatMessage.message
-                tvTime.text =
-                    formatTimeStamp(chatMessage.timestamp)
-            }
-
-            firestore.collection("User").document(chatMessage.senderId).get().addOnSuccessListener { document ->
-                val user = document.toObject(User::class.java)
-                user?.petProfile?.let {
-                    ivProfile.load(it) {
-                        error(R.drawable.kakaotalk_20230825_222509794_01)
-                    }
-                    ivProfile.setOnClickListener { userDetail(user) }
-                }
-            }
-        }
-    }
-
-
-
-    class DiffCallback : DiffUtil.ItemCallback<ChatItem>() {
+      class DiffCallback : DiffUtil.ItemCallback<ChatItem>() {
         override fun areItemsTheSame(oldItem: ChatItem, newItem: ChatItem): Boolean {
             return when {
                 oldItem is ChatItem.DateHeader && newItem is ChatItem.DateHeader -> oldItem.date == newItem.date
@@ -166,24 +101,90 @@ class ChatMessageAdapter(var context: Context) :
             return oldItem == newItem
         }
     }
-
-
-    private fun clickListener(imageView: View, imageUrl: String?) {
-        imageView.setOnClickListener {
-            val intent = Intent(it.context, ChatPullScreenActivity::class.java)
-            intent.putExtra(IMAGE_URL, imageUrl)
-            it.context.startActivity(intent)
+    inner class DateHeaderViewHolder(private val binding: ItemDateHeaderBinding) : RecyclerView.ViewHolder(binding.root) {
+        fun bind(date: String) {
+            binding.tvDateHeader.text = date
         }
     }
-
-    private fun userDetail(user: User) {
-        val intent = Intent(context, ChatClickUserDetailActivity::class.java)
-        intent.putExtra(ChatClickUserDetailActivity.USER, user)
-        context.startActivity(intent)
-    }
-    private fun formatTimeStamp(timestamp:Long):String{
-        val sdf = SimpleDateFormat("a HH:mm",Locale.KOREA)
-        sdf.timeZone = TimeZone.getTimeZone("Asia/Seoul")
-        return sdf.format(Date(timestamp))
-    }
 }
+
+//    inner class MyMessageViewHolder(private val binding: ItemMyMessageBinding) :
+//        RecyclerView.ViewHolder(binding.root) {
+//        fun bind(chatMessage: ChatMessage) = with(binding) {
+//            if (chatMessage.imageUrl.isNullOrEmpty().not()) {
+//                ivMessageImage.isVisible = true
+//                tvMessage.isVisible = false
+//                tvTime.text = formatTimeStamp(chatMessage.timestamp)
+//                ivMessageImage.load(chatMessage.imageUrl) {
+//                    error(R.drawable.kakaotalk_20230825_222509794_01)
+//                }
+//                clickListener(ivMessageImage, chatMessage.imageUrl)
+//            } else {
+//                ivMessageImage.isVisible = false
+//                tvMessage.isVisible = true
+//                tvMessage.setBackgroundResource(R.drawable.chat2)
+//                tvMessage.text = chatMessage.message
+//                tvTime.text =
+//                    formatTimeStamp(chatMessage.timestamp)
+//            }
+//        }
+//    }
+
+//    inner class OtherMessageViewHolder(private val binding: ItemOtherMessageBinding) :
+//        RecyclerView.ViewHolder(binding.root) {
+//
+//        fun bind(chatMessage: ChatMessage) = with(binding) {
+//            if (chatMessage.imageUrl.isNullOrEmpty().not()) {
+//                ivMessageImage.isVisible = true
+//                tvMessage.isVisible = false
+//                tvName.text = chatMessage.senderPetName
+//                tvTime.text = formatTimeStamp(chatMessage.timestamp)
+//                ivMessageImage.load(chatMessage.imageUrl) {
+//                    error(R.drawable.kakaotalk_20230825_222509794_01)
+//                }
+//                clickListener(ivMessageImage, chatMessage.imageUrl)
+//            } else {
+//                ivMessageImage.isVisible = false
+//                tvMessage.isVisible = true
+//                tvName.text = chatMessage.senderPetName
+//                tvMessage.setBackgroundResource(R.drawable.chat1)
+//                tvMessage.text = chatMessage.message
+//                tvTime.text =
+//                    formatTimeStamp(chatMessage.timestamp)
+//            }
+//
+//            firestore.collection("User").document(chatMessage.senderId).get().addOnSuccessListener { document ->
+//                val user = document.toObject(User::class.java)
+//                user?.petProfile?.let {
+//                    ivProfile.load(it) {
+//                        error(R.drawable.kakaotalk_20230825_222509794_01)
+//                    }
+//                    ivProfile.setOnClickListener { userDetail(user) }
+//                }
+//            }
+//        }
+//    }
+
+
+
+
+
+
+//    private fun clickListener(imageView: View, imageUrl: String?) {
+//        imageView.setOnClickListener {
+//            val intent = Intent(it.context, ChatPullScreenActivity::class.java)
+//            intent.putExtra(IMAGE_URL, imageUrl)
+//            it.context.startActivity(intent)
+//        }
+//    }
+//
+//    private fun userDetail(user: User) {
+//        val intent = Intent(context, ChatClickUserDetailActivity::class.java)
+//        intent.putExtra(ChatClickUserDetailActivity.USER, user)
+//        context.startActivity(intent)
+//    }
+//    private fun formatTimeStamp(timestamp:Long):String{
+//        val sdf = SimpleDateFormat("a HH:mm",Locale.KOREA)
+//        sdf.timeZone = TimeZone.getTimeZone("Asia/Seoul")
+//        return sdf.format(Date(timestamp))
+//    }
